factor out server into util
