+++ pkg/plugin/prommetrics.go
@@ -4,16 +4,11 @@ package plugin
 
 import (
 	"context"
-	"errors"
-	"fmt"
-	"net"
-	"net/http"
 
 	"github.com/prometheus/client_golang/prometheus"
 	"github.com/prometheus/client_golang/prometheus/collectors"
-	"github.com/prometheus/client_golang/prometheus/promhttp"
 
-	"k8s.io/klog/v2"
+	"github.com/neondatabase/autoscaling/pkg/util"
 )
 
 type PromMetrics struct {
@@ -23,13 +18,6 @@ type PromMetrics struct {
 }
 
 func (p *AutoscaleEnforcer) startPrometheusServer(ctx context.Context) error {
-	// Separate binding from serving, so that we can catch any error in this thread, rather than the
-	// server's.
-	listener, err := net.ListenTCP("tcp", &net.TCPAddr{IP: net.IPv4zero, Port: 9100})
-	if err != nil {
-		return fmt.Errorf("Error listening on TCP: %w", err)
-	}
-
 	p.metrics = PromMetrics{
 		pluginCalls: prometheus.NewCounterVec(
 			prometheus.CounterOpts{
@@ -63,28 +51,5 @@ func (p *AutoscaleEnforcer) startPrometheusServer(ctx context.Context) error {
 		p.metrics.validResourceRequests,
 	)
 
-	mux := http.NewServeMux()
-	mux.Handle("/metrics", promhttp.HandlerFor(reg, promhttp.HandlerOpts{Registry: reg}))
-	srv := &http.Server{Handler: mux}
-
-	shutdownCtx, shutdown := context.WithCancel(ctx)
-
-	// Shutdown the server when the context expires, but also exit the shutdown watcher if the
-	// server shuts down for some other reason.
-	go func() {
-		<-shutdownCtx.Done()
-		if err := srv.Shutdown(context.Background()); err != nil {
-			klog.Errorf("Error shutting down prometheus server: %w", err)
-		}
-	}()
-
-	go func() {
-		// shutdown the shutdown watcher if we exit before it
-		defer shutdown()
-		if err := srv.Serve(listener); !errors.Is(err, http.ErrServerClosed) {
-			klog.Errorf("Prometheus server exited with unexpected error: %s", err)
-		}
-	}()
-
-	return nil
+	return util.StartPrometheusMetricsServer(ctx, 9100, reg)
 }
